ms_loyalty/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pandas
openpyxl
pytest
